+++ src/belote/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "4.7.2"
+__version__ = "4.7.3"
 
 __all__ = ["__version__"]